+++ src/parse/token_list.c
@@ -6,7 +6,7 @@
 /*   By: rmarin-j <rmarin-j@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 15:24:05 by rmarin-j          #+#    #+#             */
-/*   Updated: 2025/05/13 13:34:02 by rmarin-j         ###   ########.fr       */
+/*   Updated: 2025/05/13 14:09:06 by rmarin-j         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,64 +24,6 @@ void	tk_init(t_token *new)
 	new->fd[0] = 0;
 	new->fd[1] = 1;
 	new->l_status = 0;
-}
-
-int	is_cmd(char *av, t_token *tk, t_list *env, t_data *data)
-{
-	char	**path;
-	char	*aux;
-	t_list	*key;
-	int		i;
-
-	i = 0;
-	key = find_key(env, "PATH");
-	path = ft_split(key->value, ':');
-	if (!path)
-		throw_error("ERROR: PATH not found (2)", tk, data);
-	while (path[i])
-	{
-		aux = ft_strcjoin(path[i], av, '/');
-		if (access(aux, X_OK) == 0)
-		{
-			tk->command = ft_strdup(aux);
-			tk->type = CMD;
-			free_2ptr(path);
-			free(aux);
-			return (1);
-		}
-		i++;
-	}
-	free_2ptr(path);
-	if (aux)
-		free(aux);
-	return (0);
-}
-
-int	is_builtin(t_token *tk, char *av)
-{
-	char	*aux[8];
-	int		i;
-
-	i = 0;
-	aux[0] = "export\0";
-	aux[1] = "unset\0";
-	aux[2] = "cd\0";
-	aux[3] = "pwd\0";
-	aux[4] = "env\0";
-	aux[5] = "exit\0";
-	aux[6] = "echo\0";
-	aux[7] = 0;
-	while (aux[i])
-	{
-		if (ft_strcmp(av, aux[i]) == 0)
-		{
-			tk->type = BUIL;
-			tk->command = ft_strdup(aux[i]);
-			return (1);
-		}
-		i++;
-	}
-	return (0);
 }
 
 static int	tk_argvtipe(t_token *tk_list, t_list *env, t_data *data)
@@ -106,129 +48,7 @@ static int	tk_argvtipe(t_token *tk_list, t_list *env, t_data *data)
 	tk_list->argc = i;
 	return (flag);
 }
-/* 	if (flag == 0)
-		throw_error("ERROR: no cmd in pipe\n", tk2free, NULL);
-	else if (flag > 1)
-		throw_error("ERROR: too much cmd in pipe\n", tk2free, NULL); */
 
-
-/*--------------------------------------------------------------------------------------------------------------------------------------
----------------------------------------------------------------------------------------------------------------------------------------
---------------------------------------------------------------------------------------------------------------------------------------*/
-
-int	ft_lstsize(t_list *lst)
- {
- 	t_list	*node;
- 	int		count;
-
- 	count = 0;
- 	node = lst;
- 	while (node)
- 	{
- 		count++;
- 		node = node->next;
- 	}
- 	return (count);
- }
-
-t_redir	*ft_redirlast(t_redir *rd)
-{
-	t_redir	*node;
-
-	node = rd;
-	while (node)
-	{
-		if (node->next)
-			node = node->next;
-		else
-			break ;
-	}
-	return (node);
-}
-
-void	printredir(t_redir *red, char *str)
-{
-	return ;
-	if (!red)
-		printf("la redir no existe\n");
-	printf("\n---------------\n");
-	printf("-----REDIR-----\n\n");
-	printf("PIPE = [ %s ]\n", str);
-	printf("rd--> tipo = %i,  file = %s\n", red->type, red->file);
-	printf("---------------\n\n");
-}
-int	get_redir(t_token *tk, char *str, int j, t_data *data)
-{
-	int	rd_end;
-	
-	rd_end = 0;
-	if (str[j] == '<' && str[j + 1] == '<')
-		rd_end = redir_fill(tk, str, HDOC, j, data);
-	else if (str[j] == '>' && str[j + 1] == '>')
-		rd_end = redir_fill(tk, str, NDOUT, j, data);
-	else if (str[j] == '<')
-		rd_end = redir_fill(tk, str, IN, j, data);
-	else if (str[j] == '>')
-		rd_end = redir_fill(tk, str, DOUT, j, data);
-	return(rd_end);
-}
-
-int get_av(t_list **lst, char *str, int j, t_token *tk)
-{
-	int start;
-	char *av;
-
-	start = j;
-	av = NULL;
-
-	if (str[j] == '\'' || str[j] == '\"')
-	{
-		j = end_quote(str, j + 1, str[j], tk);
-		if (j == -1)
-			return (-1);
-		av = ft_substr(str, start, j + 1 - start);
-		if (ft_strlen(av) == 2) // si hay comillas vacias pasa de ese argv
-			return (j + 1);
-		ft_lstadd_back(lst, ft_lstnew(av, "q")); //le dejo una q en el value para marcar quoted
-		return(j + 1);//aqui devuelve con comillas
-	}
-	else
-	{
-		while (!ft_isspace(str[j]) && str[j])
-			j++;
-		av = ft_substr(str, start, j - start);
-		ft_lstadd_back(lst, ft_lstnew(av, NULL));
-		return (j);
-	}
-}
-
-void print2char(char **str)
-{
-	int i = 0;
-
-	write(1, "print2char\n", 12);
-	while (str[i])
-	{
-		write(1, str[i], strlen(str[i]));
-		write(1, "\n", 1);
-		i++;
-	}
-}
-
-void print_tokenlist(t_token *tk)
-{
-	int i = 0;
-	
-	printf("\n----------TOKEN LIST----------\n");
-	while (&tk[i] && tk[i].type != NONE)
-	{
-		printf("tk[%i] :\n", i);
-		if (tk[i].argv)
-			print2char(tk[i].argv);
-		i++;
-	}
-	printf("\n----------TOKEN ENDS----------\n\n");
-}
 
 t_token	*tk_list_make(char **pipes, t_list *env, t_data *data)
 {
